+++ scripts/update_theoplayer_native.sh
@@ -22,6 +22,7 @@ update_theoplayer_ios() {
 
     sed -i '' "s/s.dependency 'THEOplayerSDK-core', '.*/s.dependency 'THEOplayerSDK-core', '${IOS}'/g" flutter_theoplayer_sdk/flutter_theoplayer_sdk_ios/ios/theoplayer_ios.podspec
     sed -i '' "s/s.dependency 'THEOplayer-Integration-THEOlive', '.*/s.dependency 'THEOplayer-Integration-THEOlive', '${IOS}'/g" flutter_theoplayer_sdk/flutter_theoplayer_sdk_ios/ios/theoplayer_ios.podspec
+    sed -i '' "s|.package(url: \"https://github.com/THEOplayer/theoplayer-sdk-apple\", exact: \".*\")|.package(url: \"https://github.com/THEOplayer/theoplayer-sdk-apple\", exact: \"${IOS}\")|g" flutter_theoplayer_sdk/flutter_theoplayer_sdk_ios/ios/theoplayer_ios/Package.swift
     pod update THEOplayerSDK-core --project-directory=flutter_theoplayer_sdk/flutter_theoplayer_sdk/example/ios
     pod update THEOplayer-Integration-THEOlive --project-directory=flutter_theoplayer_sdk/flutter_theoplayer_sdk/example/ios
     pod install --repo-update --project-directory=flutter_theoplayer_sdk/flutter_theoplayer_sdk/example/ios